Add a new DTO for Certificate validity period.
